component tests updates
