+++ features/patch_job_state_success.feature
@@ -19,7 +19,7 @@ Feature: Patch job state - Success
         
         Then the HTTP status code should be "204"
         And the response header "Content-Type" should be ""
-        And the response ETag header should be a new eTag
+        And the response ETag header should not be empty
         And the job should only be updated with the following fields and values
             | e_tag                  | new eTag                  |
             | last_updated           | now or earlier            |
@@ -43,7 +43,7 @@ Feature: Patch job state - Success
         
         And the HTTP status code should be "204"
         And the response header "Content-Type" should be ""
-        And the response ETag header should be a new eTag
+        And the response ETag header should not be empty
         And the job should only be updated with the following fields and values
             | e_tag                  | new eTag                  |
             | last_updated           | now or earlier            |
@@ -66,7 +66,7 @@ Feature: Patch job state - Success
         
         And the HTTP status code should be "204"
         And the response header "Content-Type" should be ""
-        And the response ETag header should be a new eTag
+        And the response ETag header should not be empty
         And the job should only be updated with the following fields and values
             | e_tag                  | new eTag                  |
             | last_updated           | now or earlier            |
@@ -89,7 +89,7 @@ Feature: Patch job state - Success
         
         Then the HTTP status code should be "204"
         And the response header "Content-Type" should be ""
-        And the response ETag header should be a new eTag
+        And the response ETag header should not be empty
         And the job should only be updated with the following fields and values
             | e_tag                  | new eTag                  |
             | last_updated           | now or earlier            |
@@ -113,7 +113,7 @@ Feature: Patch job state - Success
         
         Then the HTTP status code should be "204"
         And the response header "Content-Type" should be ""
-        And the response ETag header should be a new eTag
+        And the response ETag header should not be empty
         And the job should only be updated with the following fields and values
             | e_tag                  | new eTag                  |
             | last_updated           | now or earlier            |
@@ -137,7 +137,7 @@ Feature: Patch job state - Success
         
         Then the HTTP status code should be "204"
         And the response header "Content-Type" should be ""
-        And the response ETag header should be a new eTag
+        And the response ETag header should not be empty
         And the job should only be updated with the following fields and values
             | e_tag                  | new eTag                  |
             | last_updated           | now or earlier            |
@@ -161,7 +161,7 @@ Feature: Patch job state - Success
         
         Then the HTTP status code should be "204"
         And the response header "Content-Type" should be ""
-        And the response ETag header should be a new eTag
+        And the response ETag header should not be empty
         And the job should only be updated with the following fields and values
             | e_tag                  | new eTag                  |
             | last_updated           | now or earlier            |
@@ -184,7 +184,7 @@ Feature: Patch job state - Success
         
         Then the HTTP status code should be "204"
         And the response header "Content-Type" should be ""
-        And the response ETag header should be a new eTag
+        And the response ETag header should not be empty
         And the job should only be updated with the following fields and values
             | e_tag                  | new eTag                  |
             | last_updated           | now or earlier            |
