+++ features/posting_a_task.feature
@@ -1,28 +1,30 @@
-Feature: Posting a job
+Feature: Posting a task
 
   Scenario: Task is created successfully
 
     Given I use a service auth token "validServiceAuthToken"
     And zebedee recognises the service auth token as valid
+    And the search api is working correctly 
     And I have generated 1 jobs in the Job Store
     When I call POST /jobs/{id}/tasks using the generated id
     """
     { "task_name": "dataset-api", "number_of_documents": 29 }
     """
-    Then I would expect job_id, last_updated, and links to have this structure
-      | job_id       | UUID                                           |
-      | last_updated | Not in the future                              |
-      | links: self  | http://{bind_address}/jobs/{id}/tasks/dataset-api |
-      | links: job   | http://{bind_address}/jobs/{id}                |
-    And the task resource should also contain the following values:
-      | number_of_documents               | 29                        |
-      | task_name                         | dataset-api                  |
-    And the HTTP status code should be "201"
+    Then the HTTP status code should be "201"
+    And the response ETag header should not be empty
+    And the task should have the following fields and values
+      | job_id              | UUID                                              |
+      | last_updated        | Not in the future                                 |
+      | links: self         | http://{bind_address}/jobs/{id}/tasks/dataset-api |
+      | links: job          | http://{bind_address}/jobs/{id}                   |
+      | number_of_documents | 29                                                |
+      | task_name           | dataset-api                                       |
 
   Scenario: The connection to mongo DB is lost and a post request returns an internal server error
 
     Given I use a service auth token "validServiceAuthToken"
     And zebedee recognises the service auth token as valid
+    And the search api is working correctly 
     And I have generated 1 jobs in the Job Store
     And the search reindex api loses its connection to mongo DB
     When I call POST /jobs/{id}/tasks using the generated id
@@ -35,32 +37,35 @@ Feature: Posting a job
 
     Given I use a service auth token "validServiceAuthToken"
     And zebedee recognises the service auth token as valid
+    And the search api is working correctly 
     And I have generated 1 jobs in the Job Store
     And I call POST /jobs/{id}/tasks using the generated id
     """
     { "task_name": "dataset-api", "number_of_documents": 29 }
     """
     And a new task resource is created containing the following values:
-      | number_of_documents               | 29                        |
-      | task_name                         | dataset-api                  |
+      | number_of_documents               | 29                |
+      | task_name                         | dataset-api       |
     When I call POST /jobs/{id}/tasks to update the number_of_documents for that task
     """
     { "task_name": "dataset-api", "number_of_documents": 36 }
     """
-    Then I would expect job_id, last_updated, and links to have this structure
-      | job_id       | UUID                                           |
-      | last_updated | Not in the future                              |
-      | links: self  | http://{bind_address}/jobs/{id}/tasks/dataset-api |
-      | links: job   | http://{bind_address}/jobs/{id}                |
-    And the task resource should also contain the following values:
-      | number_of_documents               | 36                        |
-      | task_name                         | dataset-api                  |
-    And the HTTP status code should be "201"
+    Then the HTTP status code should be "201"
+    And the response ETag header should not be empty
+    And the task should have the following fields and values
+      | job_id              | UUID                                              |
+      | last_updated        | Not in the future                                 |
+      | links: self         | http://{bind_address}/jobs/{id}/tasks/dataset-api |
+      | links: job          | http://{bind_address}/jobs/{id}                   |
+      | number_of_documents | 36                                                |
+      | task_name           | dataset-api                                       |
+
 
   Scenario: Request body cannot be read returns a bad request error
 
     Given I use a service auth token "validServiceAuthToken"
     And zebedee recognises the service auth token as valid
+    And the search api is working correctly 
     And I have generated 1 jobs in the Job Store
     And I call POST /jobs/{id}/tasks using the generated id
     """
@@ -81,6 +86,7 @@ Feature: Posting a job
   Scenario: No authorisation header set returns a bad request error
 
     Given I am not authorised
+    And the search api is working correctly 
     And I have generated 1 jobs in the Job Store
     And I call POST /jobs/{id}/tasks using the generated id
     """
@@ -92,6 +98,7 @@ Feature: Posting a job
 
     Given I use a service auth token "invalidServiceAuthToken"
     And zebedee does not recognise the service auth token
+    And the search api is working correctly 
     And I have generated 1 jobs in the Job Store
     When I call POST /jobs/{id}/tasks using the generated id
     """
@@ -104,6 +111,7 @@ Feature: Posting a job
     Given I use an X Florence user token "validXFlorenceToken"
     And I am identified as "someone@somewhere.com"
     And zebedee recognises the user token as valid
+    And the search api is working correctly 
     And I have generated 1 jobs in the Job Store
     When I call POST /jobs/{id}/tasks using the generated id
     """
@@ -116,6 +124,7 @@ Feature: Posting a job
     Given I use an X Florence user token "invalidXFlorenceToken"
     And I am not identified by zebedee
     And zebedee does not recognise the user token
+    And the search api is working correctly 
     And I have generated 1 jobs in the Job Store
     When I call POST /jobs/{id}/tasks using the generated id
     """
@@ -127,6 +136,7 @@ Feature: Posting a job
 
     Given I use a service auth token "validServiceAuthToken"
     And zebedee recognises the service auth token as valid
+    And the search api is working correctly 
     And I have generated 1 jobs in the Job Store
     And I call POST /jobs/{id}/tasks using the generated id
     """
